requirements.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.10.2
django-hosts==2.0
django-taggit==0.22.1
djangorestframework==3.6.3
gunicorn==19.7.1
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
six==1.10.0
whitenoise==2.0.6
